requirements.txt
arxiv
semanticscholar
scholarly
habanero
unpywall
sickle
requests
beautifulsoup4
pandas
google-generativeai
tqdm
python-dotenv
streamlit
google-auth-oauthlib
google-api-python-client
